+++ Node.py
@@ -1,0 +1,5 @@
+class Node:
+    def __init__(self, dados=None):
+        self.dados = dados
+        self.anterior = None
+        self.proximo = None
